requirements.txt
#env 
python-dotenv

# web scraping 
selenium
undetected-chromedriver
scrapy
